divide a matrix

+++ python-test_driven_development/2-matrix_divided.py
@@ -24,6 +24,6 @@ def matrix_divided(matrix, div):
             if type(s) not in [int, float]:
                 raise TypeError("matrix must
                                 be a matrix(list of lists) of integers/floats")
-            new_row.append(int(s / div))
+            new_row.append(s / div))
         new_matrix.append(new_row)
     return new_matrix
